+++ gymnasium_robotics/__init__.py
@@ -21,63 +21,63 @@ def register_robotics_envs():
         # Fetch
         register(
             id=f"FetchSlide{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoPyFetchSlideEnv",
+            entry_point="gymnasium_robotics.envs.fetch.slide:MujocoPyFetchSlideEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"FetchSlide{suffix}-v2",
-            entry_point="gymnasium_robotics.envs:MujocoFetchSlideEnv",
+            entry_point="gymnasium_robotics.envs.fetch.slide:MujocoFetchSlideEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"FetchPickAndPlace{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoPyFetchPickAndPlaceEnv",
+            entry_point="gymnasium_robotics.envs.fetch.pick_and_place:MujocoPyFetchPickAndPlaceEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"FetchPickAndPlace{suffix}-v2",
-            entry_point="gymnasium_robotics.envs:MujocoFetchPickAndPlaceEnv",
+            entry_point="gymnasium_robotics.envs.fetch.pick_and_place:MujocoFetchPickAndPlaceEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"FetchReach{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoPyFetchReachEnv",
+            entry_point="gymnasium_robotics.envs.fetch.reach:MujocoPyFetchReachEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"FetchReach{suffix}-v2",
-            entry_point="gymnasium_robotics.envs:MujocoPyFetchReachEnv",
+            entry_point="gymnasium_robotics.envs.fetch.reach:MujocoPyFetchReachEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"FetchReach{suffix}-v3",
-            entry_point="gymnasium_robotics.envs:MujocoFetchReachEnv",
+            entry_point="gymnasium_robotics.envs.fetch.reach:MujocoFetchReachEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"FetchPush{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoPyFetchPushEnv",
+            entry_point="gymnasium_robotics.envs.fetch.push:MujocoPyFetchPushEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"FetchPush{suffix}-v2",
-            entry_point="gymnasium_robotics.envs:MujocoFetchPushEnv",
+            entry_point="gymnasium_robotics.envs.fetch.push:MujocoFetchPushEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
@@ -85,21 +85,21 @@ def register_robotics_envs():
         # Hand
         register(
             id=f"HandReach{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandReachEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.reach:MujocoPyHandReachEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"HandReach{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandReachEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.reach:MujocoHandReachEnv",
             kwargs=kwargs,
             max_episode_steps=50,
         )
 
         register(
             id=f"HandManipulateBlockRotateZ{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoPyHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -112,7 +112,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateZ{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -125,7 +125,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateZ_BooleanTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoPyHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -139,7 +139,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateZ_BooleanTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -153,7 +153,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateZ_ContinuousTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoPyHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -167,7 +167,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateZ_ContinuousTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -181,7 +181,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateParallel{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoPyHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -194,7 +194,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateParallel{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -207,7 +207,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateParallel_BooleanTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoPyHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -221,7 +221,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateParallel_BooleanTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -235,7 +235,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateParallel_ContinuousTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoPyHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -249,7 +249,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateParallel_ContinuousTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -263,7 +263,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateXYZ{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoPyHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -276,7 +276,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateXYZ{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -289,7 +289,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateXYZ_BooleanTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoPyHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -303,7 +303,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateXYZ_BooleanTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -317,7 +317,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateXYZ_ContinuousTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoPyHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -331,7 +331,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockRotateXYZ_ContinuousTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -345,7 +345,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockFull{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoPyHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -358,7 +358,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlockFull{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -372,7 +372,7 @@ def register_robotics_envs():
         # Alias for "Full"
         register(
             id=f"HandManipulateBlock{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoPyHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -385,7 +385,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlock{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block:MujocoHandBlockEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -398,7 +398,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlock_BooleanTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoPyHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -412,7 +412,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlock_BooleanTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -426,7 +426,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlock_ContinuousTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoPyHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -440,7 +440,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateBlock_ContinuousTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandBlockTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_block_touch_sensors:MujocoHandBlockTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -454,7 +454,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEggRotate{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandEggEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg:MujocoPyHandEggEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -467,7 +467,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEggRotate{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandEggEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg:MujocoHandEggEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -480,7 +480,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEggRotate_BooleanTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandEggTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg_touch_sensors:MujocoPyHandEggTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -494,7 +494,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEggRotate_BooleanTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandEggTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg_touch_sensors:MujocoHandEggTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -508,7 +508,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEggRotate_ContinuousTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandEggTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg_touch_sensors:MujocoPyHandEggTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -522,7 +522,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEggRotate_ContinuousTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandEggTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg_touch_sensors:MujocoHandEggTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -536,7 +536,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEggFull{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandEggEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg:MujocoPyHandEggEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -549,7 +549,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEggFull{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandEggEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg:MujocoHandEggEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -563,7 +563,7 @@ def register_robotics_envs():
         # Alias for "Full"
         register(
             id=f"HandManipulateEgg{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandEggEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg:MujocoPyHandEggEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -576,7 +576,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEgg{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandEggEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg:MujocoHandEggEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -589,7 +589,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEgg_BooleanTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandEggTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg_touch_sensors:MujocoPyHandEggTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -603,7 +603,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEgg_BooleanTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandEggTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg_touch_sensors:MujocoHandEggTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -617,7 +617,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEgg_ContinuousTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandEggTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg_touch_sensors:MujocoPyHandEggTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -631,7 +631,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulateEgg_ContinuousTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandEggTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_egg_touch_sensors:MujocoHandEggTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -645,7 +645,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePenRotate{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandPenEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen:MujocoPyHandPenEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -658,7 +658,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePenRotate{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandPenEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen:MujocoHandPenEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -671,7 +671,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePenRotate_BooleanTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandPenTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen_touch_sensors:MujocoPyHandPenTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -685,7 +685,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePenRotate_BooleanTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandPenTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen_touch_sensors:MujocoHandPenTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -699,7 +699,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePenRotate_ContinuousTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandPenTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen_touch_sensors:MujocoPyHandPenTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -713,7 +713,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePenRotate_ContinuousTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandPenTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen_touch_sensors:MujocoHandPenTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "ignore",
@@ -727,7 +727,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePenFull{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandPenEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen:MujocoPyHandPenEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -740,7 +740,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePenFull{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandPenEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen:MujocoHandPenEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -753,7 +753,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePen{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandPenEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen:MujocoPyHandPenEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -766,7 +766,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePen{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandPenEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen:MujocoHandPenEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -779,7 +779,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePen_BooleanTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandPenTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen_touch_sensors:MujocoPyHandPenTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -793,7 +793,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePen_BooleanTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandPenTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen_touch_sensors:MujocoHandPenTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -807,7 +807,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePen_ContinuousTouchSensors{suffix}-v0",
-            entry_point="gymnasium_robotics.envs:MujocoPyHandPenTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen_touch_sensors:MujocoPyHandPenTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
@@ -821,7 +821,7 @@ def register_robotics_envs():
 
         register(
             id=f"HandManipulatePen_ContinuousTouchSensors{suffix}-v1",
-            entry_point="gymnasium_robotics.envs:MujocoHandPenTouchSensorsEnv",
+            entry_point="gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_pen_touch_sensors:MujocoHandPenTouchSensorsEnv",
             kwargs=_merge(
                 {
                     "target_position": "random",
