--- gymnasium_robotics/envs/__init__.py
@@ -1,31 +1,0 @@
-# noqa: D104
-from gymnasium_robotics.envs.fetch.pick_and_place import (
-    MujocoFetchPickAndPlaceEnv,
-    MujocoPyFetchPickAndPlaceEnv,
-)
-from gymnasium_robotics.envs.fetch.push import MujocoFetchPushEnv, MujocoPyFetchPushEnv
-from gymnasium_robotics.envs.fetch.reach import (
-    MujocoFetchReachEnv,
-    MujocoPyFetchReachEnv,
-)
-from gymnasium_robotics.envs.fetch.slide import (
-    MujocoFetchSlideEnv,
-    MujocoPyFetchSlideEnv,
-)
-from gymnasium_robotics.envs.hand.manipulate import (
-    MujocoHandBlockEnv,
-    MujocoHandEggEnv,
-    MujocoHandPenEnv,
-    MujocoPyHandBlockEnv,
-    MujocoPyHandEggEnv,
-    MujocoPyHandPenEnv,
-)
-from gymnasium_robotics.envs.hand.manipulate_touch_sensors import (
-    MujocoHandBlockTouchSensorsEnv,
-    MujocoHandEggTouchSensorsEnv,
-    MujocoHandPenTouchSensorsEnv,
-    MujocoPyHandBlockTouchSensorsEnv,
-    MujocoPyHandEggTouchSensorsEnv,
-    MujocoPyHandPenTouchSensorsEnv,
-)
-from gymnasium_robotics.envs.hand.reach import MujocoHandReachEnv, MujocoPyHandReachEnv
