+++ gymnasium_robotics/envs/fetch/__init__.py
@@ -1,0 +1,1 @@
+from gymnasium_robotics.envs.fetch.fetch_env import MujocoFetchEnv, MujocoPyFetchEnv
