+++ gymnasium_robotics/envs/shadow_dexterous_hand/__init__.py
@@ -1,0 +1,12 @@
+from gymnasium_robotics.envs.shadow_dexterous_hand.hand_env import (
+    MujocoHandEnv,
+    MujocoPyHandEnv,
+)
+from gymnasium_robotics.envs.shadow_dexterous_hand.manipulate import (
+    MujocoManipulateEnv,
+    MujocoPyManipulateEnv,
+)
+from gymnasium_robotics.envs.shadow_dexterous_hand.manipulate_touch_sensors import (
+    MujocoManipulateTouchSensorsEnv,
+    MujocoPyManipulateTouchSensorsEnv,
+)
